model.e2e: first implementation
